requirements.txt
datetime
matplotlib
numpy
pandas
pyyaml
python-dotenv
requests
mysql-connector-python
